move client registration into its own file
