+++ src/phoenix.js
@@ -45,6 +45,8 @@ import Avatar from './components/Avatar.vue'
 
 import wgxpath from 'wicked-good-xpath'
 
+import { registerClient } from './services/clientRegistration'
+
 wgxpath.install()
 
 Vue.prototype.$client = new OwnCloud()
@@ -219,61 +221,6 @@ function missingConfig () {
     render: h => h(missingConfigPage)
   })
 }
-
-async function get(url){
-  return (await (fetch(url)
-      .then(res => {
-        return res.json()
-      })
-      .catch(err => {
-        console.log('Error: ', err)
-      })
-  ))
-}
-
-async function post(url, data) {
-  return (await (fetch(url, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(data),
-    })
-      .then(res => {
-        return res.json()
-      })
-      .catch(err => {
-        console.log('Error: ', err)
-      })
-  ))
-}
-
-async function registerClient(openIdConfig) {
-  const clientData = sessionStorage.getItem('dynamicClientData')
-  if (clientData !== null) {
-    const client_secret_expires_at = clientData.client_secret_expires_at || 0;
-    if (client_secret_expires_at === 0 || Date.now() < client_secret_expires_at*1000) {
-      return JSON.parse(clientData)
-    }
-  }
-  sessionStorage.removeItem('dynamicClientData')
-
-  let baseUrl = window.location.href.split('#')[0]
-  if (baseUrl.endsWith('/index.html')) {
-    baseUrl = baseUrl.substr(0, baseUrl.length - 10)
-  }
-
-  const wellKnown = await get(`${openIdConfig.authority}/.well-known/openid-configuration`)
-  const resp = await post(wellKnown.registration_endpoint, {
-    "redirect_uris": [
-      baseUrl + 'oidc-callback.html'
-    ],
-    "client_name": `ownCloud Phoenix(${appVersionJson.version}) on ${baseUrl}`
-  })
-  sessionStorage.setItem('dynamicClientData', JSON.stringify(resp))
-  return resp
-}
-
 
 (async function () {
   config = await fetch('config.json')
